Moved checks with badges to their own file badge_checks.py
